refactor(chat-highlight): extract pure helpers for testability
Moves isCreditError, extractPendingApprovals, and extractPendingPlans
into dedicated .ts modules so they can be imported by bun:test code
(use-highlight-count.test.ts) without dragging in @deco/ui transitively.
Original .tsx callers keep working via re-exports. Removes the inlined
duplicates introduced in f010ae78c.
