+++ apps/mesh/src/web/components/chat/highlight/approval.tsx
@@ -17,23 +17,14 @@ import {
   usePreferences,
   type ToolApprovalLevel,
 } from "@/web/hooks/use-preferences.ts";
-import { stripMcpServerPrefix } from "@/web/lib/tool-namespace";
-import { toTitleCase } from "../message/parts/tool-call-part/utils.tsx";
 import { CollapsibleHighlight } from "./collapsible-highlight";
 import { PaginatedFormFooterLeft } from "./common/paginated-form-footer";
 import { useMultiPartDecisionForm } from "./common/use-multipart-decision-form";
-
-// ============================================================================
-// Types
-// ============================================================================
-
-export interface PendingApproval {
-  approvalId: string;
-  toolCallId: string;
-  toolName: string;
-  friendlyName: string;
-  input: unknown;
-}
+import { type PendingApproval } from "./extract-pending-approvals";
+export {
+  extractPendingApprovals,
+  type PendingApproval,
+} from "./extract-pending-approvals";
 
 // ============================================================================
 // Constants
@@ -416,46 +407,4 @@ export function ApprovalHighlight({
       onRespond={onRespond}
     />
   );
-}
-
-// ============================================================================
-// Utility: extract pending approvals from message parts
-// ============================================================================
-
-export function extractPendingApprovals(
-  parts: Array<{
-    type: string;
-    state?: string;
-    approval?: { id: string };
-    toolCallId?: string;
-    toolName?: string;
-    input?: unknown;
-  }>,
-): PendingApproval[] {
-  const result: PendingApproval[] = [];
-  for (const part of parts) {
-    if (
-      "state" in part &&
-      part.state === "approval-requested" &&
-      "approval" in part &&
-      part.approval?.id &&
-      "toolCallId" in part &&
-      part.toolCallId
-    ) {
-      const toolName =
-        "toolName" in part && typeof part.toolName === "string"
-          ? part.toolName
-          : part.type.startsWith("tool-")
-            ? part.type.replace("tool-", "")
-            : "Tool";
-      result.push({
-        approvalId: part.approval.id,
-        toolCallId: part.toolCallId,
-        toolName,
-        friendlyName: toTitleCase(stripMcpServerPrefix(toolName)),
-        input: part.input,
-      });
-    }
-  }
-  return result;
 }
